Store refs to TownBlocks in the TownBlock table
Storing references to townblocks in a massive string is a HUGE waste of
time. With large towns it freezes the main thread for very long periods
of time when they are required to update. (Just from building that
string)

Instead, storing a reference to the town each town block belongs to
inside the townblocks table allows updating and loading of town blocks
to occur without issue.